Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main(){
-    int i,j,k;
+    float i,j,k;
     float avg;
-    scanf("%d %d %d",&i,&j,&k);
+    scanf("%f %f %f",&i,&j,&k);
     avg= (i+j+k)/3;
     printf("Average: %.2f",avg);
     return 0;
